replay viewer: add 1/2x playback speed
The replay transport gains a 0.5x speed: command '5' selects a replay-only
half speed (ReplayHalfSpeedIndex) that spends one sim tick every other frame
(halfPhase parity, outside the lull boost), the chrome grows a 0.5x chip, and
the frame JSON's sp field carries 0.5 while it is active (buildStateJson speed
is a float now). '-' floors on 1/2x instead of 1x.

Space already pauses on the one shipped page (client/replay_broadcast.html's
keydown, which the Dockerfile splices into dist/index.html); there is no
iframe shell in this repo, so nothing needs to forward it. Test 58 pins that
binding, the digit-key relay that carries '5', and the engine's speed list.

chrome_common.js keeps reading window.CTF_WIRE — tools/gen_wire_constants.nim
emits that name as an alias of MINIGRID_WIRE by design — so the only edit to
the starter's chrome is the 0.5x fallback and the 0.5x -> '5' command map.
The viewer test re-pins those bytes (40037, sha 594ed4a7…) and now spells out
both patched lines so a wider edit cannot hide behind a re-pinned hash.

Verified: nim c -d:release src/minigrid.nim; nim check -d:emscripten
replay-viewer/minigrid_replay.nim; all 12 tests/*.nim in debug and -d:release;
tune_baselines --check; and a real emcc bundle in headless chromium —
viewer_smoke (loaded, 10s soak, distinct scrub readouts, no 404s) plus a probe
that measured 72 ticks/3s at 1x vs 36 ticks/3s at 0.5x, the 0.5x chip
highlighting, and Space pausing then resuming.

+++ src/minigrid/broadcast.nim
@@ -295,7 +295,7 @@ proc buildStateJson*(
   sim: SimServer,
   events: JsonNode,
   playing: bool,
-  speed: int,
+  speed: float,
   maxTick: int,
   looping: bool,
   transportEnabled: bool,
